Fixed warnings, improved compatability.

+++ include/setup.h
@@ -255,7 +255,7 @@ struct Time
     double framerate()
     {
         double sum = 0;
-        for(int i=0; i<framerates.size(); i++)
+        for(size_t i=0; i<framerates.size(); i++)
         {
             sum += framerates[i];
         }
